# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "gvp是语音平台的后台管理接口服务器",
        "title": "GVP API",
        "contact": {},
        "license": {},
        "version": "0.0.1"
    },
    "paths": {
        "/api/v1/bind": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "获取多个绑定关系",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "游戏名称",
                        "name": "gameid",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "游戏名称",
                        "name": "serviceid",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            },
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "添加绑定关系",
                "parameters": [
                    {
                        "maxLength": 100,
                        "type": "string",
                        "description": "游戏名称",
                        "name": "game",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "服务ID",
                        "name": "serviceid",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "游戏ID",
                        "name": "gameid",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/bind/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "获取单个绑定信息",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "summary": "删除绑定关系",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "绑定关系ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/client/svc": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "获取服务商信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "游戏用户appid",
                        "name": "appid",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "游戏用户appkey",
                        "name": "appkey",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "服务商名称",
                        "name": "svcname",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "app.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "object"
                },
                "msg": {
                    "type": "string"
                }
            }
        }
    }
}